server/requirements.txt
flask
python-dotenv
psycopg2-binary
flask-cors
PyJWT